Changed layout of index page and started on individual club pages
Instead of the accordion layout which is on the site now, I have changed
it to a more simple dropdown menu interface. The students and teachers I
asked preferred my design.
Individual club pages are not quite done yet, I still need information
for most of them. All the information on the pages now was taken from
the mini.html and the original index, so not everything has all of the
necessary information, such as room numbers or meeting dates/times.