Added content, fixed bugs, polished layout, and more.
Added content to all of the club pages that i got any information for,
however some clubs are still missing some or all of the information that
I wanted.
Also fixed some bugs which I found when the page was displayed at lower
resolutions.
Finally, I also made the second header an SSI for quicker editing.