Removed the Layout file for the Sports index page
Rails has the option to render pages from the Controller with a template of sorts, so that you can limit redundancies within files.  However, to make things simpler to edit and such, we're probably better of only using these sometimes.  While I originally had the sports index page rendering within it's own template file, I realized that this was ridiculously redundant and instead just condensed it all to one file.