Sports Games Working
As it turns out, the only issue with the games was that the Time did not
like to be updated through the Admin Interface. As the time is probably
the least necessary of the fields, and nowhere near worth the time it
has taken me to troubleshoot the issue, I removed it to get the Sports
Games interface functional.